requirements.txt
# Core LangChain / RAG
langgraph==0.2.59
langchain-community
langchain-chroma==0.1.4
langchain-ollama==0.2.1
langchain-huggingface==0.1.2

# PDF loading (compatível com Python 3.12)
pdfminer.six==20221105
unstructured[pdf]==0.14.6

posthog>=2.4.0,<6.0.0
raglight
